Fallback implementation of os.clock for windows
As windows doesn't have syscall.Getrusage
